+++ packages/core-web/src/styles/markbind.css
@@ -142,6 +142,18 @@ footer {
     color: #777;
 }
 
+.large {
+    font-size: 125%;
+}
+
+.small {
+    font-size: 75%;
+}
+
+.underline {
+    text-decoration: underline;
+}
+
 /* Bootstrap small(sm) responsive breakpoint */
 @media (max-width: 767.98px) {
     .dropdown-menu > li > a {
